refactor(node): record recent transactions through a TransactionLogger port
Decouple the indexer's transaction processor from the debug-page backend.
The processor no longer takes an mpsc::Sender<SubmittedTransaction> and no
longer knows about the channel: it takes an infallible TransactionLogger and
calls log_transaction, so full/closed handling lives behind the port.

Move the shared record types (SubmittedTransaction, SubmittedTransactionStatus,
SubmittedTxMetadata, SignerContext) and the TransactionLogger trait into the
neutral crate::types module, so neither the indexer nor the web module depends
on the other. The web module keeps storage/render plus RecentTransactionsLogger
(the channel-backed port impl, warn-on-drop) and spawn_recent_transactions_drain.

Wire the channel, buffer, and drain task in run.rs and pass the shared buffer to
the web server, so the drain is no longer spawned inside start_web_server.
