+++ crates/node/src/run.rs
@@ -14,9 +14,13 @@ use crate::{
     profiler,
     tracing::init_logging,
     tracking::{self, start_root_task},
+    types::SubmittedTransaction,
     web::{
         DebugRequest,
-        recent_transactions::{RECENT_TRANSACTIONS_CHANNEL_SIZE, SubmittedTransaction},
+        recent_transactions::{
+            RECENT_TRANSACTIONS_CHANNEL_SIZE, RecentTransactionsLogger, SharedRecentTransactions,
+            spawn_recent_transactions_drain,
+        },
         start_web_server, static_web_data,
     },
 };
@@ -173,12 +177,14 @@ pub async fn run_mpc_node(config: StartConfig) -> anyhow::Result<()> {
 
     let (migration_state_sender, migration_state_receiver) = watch::channel((0, BTreeMap::new()));
 
-    // Channel carrying completed recent-transaction records, surfaced on
-    // `/debug/recent_transactions`. The sender goes to the indexer's
-    // transaction processor (spawned below); the receiver is drained into the
-    // buffer owned by the web server (started below).
+    // Buffer behind the recent-transactions debug page. The indexer forwards
+    // records over `recent_tx_sender`; the drain task records them into the
+    // buffer, which the web server reads for snapshots. The drain is aborted on
+    // shutdown.
     let (recent_tx_sender, recent_tx_receiver) =
         mpsc::channel::<SubmittedTransaction>(RECENT_TRANSACTIONS_CHANNEL_SIZE);
+    let recent_transactions = SharedRecentTransactions::default();
+    spawn_recent_transactions_drain(recent_tx_receiver, recent_transactions.clone());
 
     let web_server = root_runtime
         .block_on(start_web_server(
@@ -189,7 +195,7 @@ pub async fn run_mpc_node(config: StartConfig) -> anyhow::Result<()> {
             protocol_state_receiver,
             migration_state_receiver,
             config.node.clone(),
-            recent_tx_receiver,
+            recent_transactions.clone(),
         ))
         .context("Failed to create web server.")?;
 
@@ -208,7 +214,7 @@ pub async fn run_mpc_node(config: StartConfig) -> anyhow::Result<()> {
         migration_state_sender,
         *tls_public_key,
         node_config.foreign_chains.clone(),
-        recent_tx_sender,
+        RecentTransactionsLogger::new(recent_tx_sender),
     );
 
     let cancellation_token = CancellationToken::new();
