+++ crates/node/src/web/recent_transactions.rs
@@ -1,5 +1,6 @@
-//! In-memory record of recently submitted transactions, shown on the
-//! `/debug/recent_transactions` web page.
+//! In-memory storage and rendering for the `/debug/recent_transactions` web
+//! page. The record types live in `crate::types`, so neither the indexer nor
+//! this module depends on the other.
 //!
 //! A successful submission only means the RPC accepted the transaction, not
 //! that it was included in a block or had its intended effect. That effect is
@@ -8,144 +9,30 @@
 //! record (txid, nonce, signer access key, ...) so an operator can debug
 //! failures such as out-of-order nonce rejections.
 
-use near_account_id::AccountId;
-use near_crypto::Signature;
-use near_indexer_primitives::{
-    CryptoHash,
-    types::{BlockHeight, Nonce},
-};
-use near_mpc_contract_interface::types::Ed25519PublicKey;
-use near_time::Utc;
+use crate::types::{SubmittedTransaction, TransactionLogger};
 use std::collections::VecDeque;
 use std::fmt::Write;
 use std::sync::{Arc, Mutex};
+use tokio::sync::mpsc;
 
 /// The most recent submitted transactions to retain; older entries are evicted
 /// once the buffer is full.
 const NUM_RECENT_TRANSACTIONS_TO_KEEP: usize = 2000;
 
-/// Capacity of the channel that carries [`SubmittedTransaction`] records from
-/// the indexer's transaction processor to the web server's drain task.
+/// Capacity of the channel carrying [`SubmittedTransaction`] records from the
+/// indexer's transaction processor to the drain task spawned in `run.rs`. One
+/// record per submitted transaction, so a full channel (a dropped record) is
+/// not expected in practice; sized to cap worst-case memory.
 pub const RECENT_TRANSACTIONS_CHANNEL_SIZE: usize = 10000;
-
-/// The observed lifecycle outcome of a submitted transaction.
-#[derive(Clone, Copy, Debug, PartialEq, Eq)]
-pub enum SubmittedTransactionStatus {
-    /// Building, signing, or routing the transaction failed locally; it never
-    /// reached the network.
-    SubmitFailed,
-    /// The transaction's intended effect was observed on chain.
-    Executed,
-    /// The transaction's intended effect was not observed before the timeout
-    /// (it may have been rejected, e.g. for a stale nonce, or simply delayed).
-    NotExecuted,
-    /// The transaction type has no on-chain effect we verify, so the outcome is
-    /// not determinable from contract state.
-    Unknown,
-    /// An error occurred while trying to observe the on-chain effect.
-    ObserveError,
-}
-
-/// A single submitted transaction and its final status.
-#[derive(Clone, Debug, PartialEq, Eq, derive_more::Display)]
-#[display(
-    "  {submitted_at}  {:<12}  method={method:<24}  signer={signer_account_id} key={}  {}",
-    format!("{status:?}"),
-    String::from(signer_public_key),
-    metadata
-        .as_ref()
-        .map_or_else(|| "(not submitted)".to_string(), |m| m.to_string())
-)]
-pub struct SubmittedTransaction {
-    /// The built-and-signed transaction details (txid, nonce, signature, block
-    /// height). Absent if building/signing failed before they were produced;
-    /// present otherwise, all together.
-    pub metadata: Option<SubmittedTxMetadata>,
-    /// The account the transaction was submitted from.
-    pub signer_account_id: AccountId,
-    /// The access key (public key) the transaction was signed with.
-    pub signer_public_key: Ed25519PublicKey,
-    /// The contract method invoked (e.g. `respond`, `respond_ckd`).
-    pub method: &'static str,
-    /// Wall-clock time at which the transaction was routed (or, on the
-    /// submit-failed path, at which the failure was recorded). Captured at
-    /// submission time, not when the outcome was later observed.
-    pub submitted_at: Utc,
-    /// The final observed outcome.
-    pub status: SubmittedTransactionStatus,
-}
-
-/// The signer-specific context for a submission, known before the transaction
-/// is built and shared by every record produced for that submission.
-pub struct SignerContext {
-    pub account_id: AccountId,
-    pub public_key: Ed25519PublicKey,
-    pub method: &'static str,
-}
-
-/// The metadata of a successfully built-and-submitted transaction, captured in
-/// `crate::indexer::tx_sender`.
-#[derive(Clone, Debug, PartialEq, Eq, derive_more::Display)]
-#[display("txid={tx_hash}  nonce={nonce}  block={block_height}  sig={signature}")]
-pub struct SubmittedTxMetadata {
-    pub tx_hash: CryptoHash,
-    pub nonce: Nonce,
-    pub signature: Signature,
-    pub block_height: BlockHeight,
-}
-
-impl SubmittedTransaction {
-    /// A record for a transaction that was successfully built and routed, with
-    /// its observed on-chain [`SubmittedTransactionStatus`]. `submitted_at` is
-    /// the time the transaction was routed, captured by the caller before it
-    /// waited to observe the outcome.
-    pub fn submitted(
-        signer: SignerContext,
-        metadata: SubmittedTxMetadata,
-        status: SubmittedTransactionStatus,
-        submitted_at: Utc,
-    ) -> Self {
-        Self::new(signer, Some(metadata), status, submitted_at)
-    }
-
-    /// A record for a transaction that could not be built, signed, or routed and
-    /// so never reached the network. `submitted_at` is the time the failure was
-    /// recorded.
-    pub fn submit_failed(signer: SignerContext, submitted_at: Utc) -> Self {
-        Self::new(
-            signer,
-            None,
-            SubmittedTransactionStatus::SubmitFailed,
-            submitted_at,
-        )
-    }
-
-    /// Builds a record from the signer context and, when the transaction was
-    /// successfully built, its metadata.
-    fn new(
-        signer: SignerContext,
-        metadata: Option<SubmittedTxMetadata>,
-        status: SubmittedTransactionStatus,
-        submitted_at: Utc,
-    ) -> Self {
-        Self {
-            metadata,
-            signer_account_id: signer.account_id,
-            signer_public_key: signer.public_key,
-            method: signer.method,
-            submitted_at,
-            status,
-        }
-    }
-}
 
 /// A bounded log of recently submitted transactions, one row per submission.
 /// Newest entries are at the back; the oldest are evicted once the buffer is
 /// full.
 ///
 /// A row is recorded once, after its outcome is observed (or after a local
-/// submit failure), so every row carries its final [`SubmittedTransactionStatus`]
-/// and rows are never mutated in place.
+/// submit failure), so every row carries its final
+/// [`SubmittedTransactionStatus`](crate::types::SubmittedTransactionStatus) and
+/// rows are never mutated in place.
 #[derive(Default)]
 pub struct RecentTransactions {
     /// One row per submission, front oldest / back newest.
@@ -174,9 +61,11 @@ impl RecentTransactions {
     }
 }
 
-/// Shared handle to a [`RecentTransactions`] buffer that the web server's drain
-/// task writes to and the request handler reads from, with the synchronization
-/// encapsulated so callers just `record` and `snapshot`.
+/// Shared handle to a [`RecentTransactions`] buffer. It is fed over a bounded
+/// channel by the indexer's [`RecentTransactionsLogger`], drained into it by the
+/// task [`spawn_recent_transactions_drain`] starts in `run.rs`, and read by the
+/// request handler, with the synchronization encapsulated so callers just
+/// `record` and `snapshot`.
 #[derive(Clone, Default)]
 pub struct SharedRecentTransactions(Arc<Mutex<RecentTransactions>>);
 
@@ -195,6 +84,61 @@ impl SharedRecentTransactions {
     pub fn snapshot(&self) -> Vec<SubmittedTransaction> {
         self.0.lock().expect("lock must not be poisoned").snapshot()
     }
+}
+
+/// [`TransactionLogger`] that forwards records over a bounded channel to the
+/// drain task spawned in `run.rs`.
+#[derive(Clone)]
+pub struct RecentTransactionsLogger {
+    sender: mpsc::Sender<SubmittedTransaction>,
+}
+
+impl RecentTransactionsLogger {
+    /// Wraps the producing end of the recent-transactions channel; the
+    /// consuming end is drained by [`spawn_recent_transactions_drain`].
+    pub fn new(sender: mpsc::Sender<SubmittedTransaction>) -> Self {
+        Self { sender }
+    }
+}
+
+impl TransactionLogger for RecentTransactionsLogger {
+    fn log_transaction(&self, transaction: SubmittedTransaction) {
+        match self.sender.try_send(transaction) {
+            Ok(()) => {}
+            Err(mpsc::error::TrySendError::Full(_)) => {
+                tracing::warn!(
+                    target: "mpc",
+                    "recent-transactions channel full; dropping debug record"
+                );
+            }
+            Err(mpsc::error::TrySendError::Closed(_)) => {
+                tracing::warn!(
+                    target: "mpc",
+                    "recent-transactions drain task gone; dropping debug record"
+                );
+            }
+        }
+    }
+}
+
+/// Spawns the task that drains the recent-transactions channel into `buffer`,
+/// recording each record the indexer's [`RecentTransactionsLogger`] forwards.
+/// The loop ends if every sender drops (`recv` returns `None`); at shutdown the
+/// task is usually just aborted when the runtime is dropped. Must be called
+/// within a Tokio runtime.
+pub fn spawn_recent_transactions_drain(
+    mut receiver: mpsc::Receiver<SubmittedTransaction>,
+    buffer: SharedRecentTransactions,
+) {
+    tokio::spawn(async move {
+        while let Some(transaction) = receiver.recv().await {
+            buffer.record(transaction);
+        }
+        tracing::warn!(
+            target: "mpc",
+            "recent-transactions channel closed; drain task exiting"
+        );
+    });
 }
 
 /// Renders a snapshot (newest first, as returned by
@@ -220,15 +164,13 @@ pub fn render(transactions: &[SubmittedTransaction]) -> String {
 #[expect(non_snake_case)]
 mod tests {
     use super::*;
+    use crate::types::{SubmittedTransactionStatus, SubmittedTxMetadata};
+    use near_account_id::AccountId;
+    use near_crypto::Signature;
+    use near_indexer_primitives::CryptoHash;
+    use near_mpc_contract_interface::types::Ed25519PublicKey;
+    use near_time::Utc;
     use std::str::FromStr;
-
-    /// A submitted (Executed) transaction whose hash is fixed
-    /// (`CryptoHash::default()`). Used by the `Display` tests, which pin the
-    /// exact rendered txid. Buffer tests that keep entries distinct should use
-    /// [`test_transaction_with_hash`].
-    fn test_transaction(method: &'static str) -> SubmittedTransaction {
-        test_transaction_with_hash(method, CryptoHash::default())
-    }
 
     /// A submitted (Executed) transaction with the given method and hash.
     fn test_transaction_with_hash(
@@ -315,7 +257,7 @@ mod tests {
         let failed = SubmittedTransaction {
             metadata: None,
             status: SubmittedTransactionStatus::SubmitFailed,
-            ..test_transaction("failed")
+            ..test_transaction_with_hash("failed", CryptoHash::default())
         };
         let submitted = test_transaction_with_hash("respond", hash(1));
         let mut buffer = RecentTransactions::default();
@@ -326,36 +268,6 @@ mod tests {
 
         // Then the failed row appears in the snapshot, newest first.
         assert_eq!(buffer.snapshot(), vec![failed, submitted]);
-    }
-
-    #[test]
-    fn submitted_transaction_display__should_render_exact_line_with_padding() {
-        // Given
-        let transaction = test_transaction("respond");
-
-        // When
-        let rendered = transaction.to_string();
-
-        // Then
-        let expected = "  2023-11-14 22:13:20.0 +00:00:00  Executed      method=respond                   signer=responder.near key=ed25519:US517G5965aydkZ46HS38QLi7UQiSojurfbQfKCELFx  txid=11111111111111111111111111111111  nonce=7  block=42  sig=ed25519:1111111111111111111111111111111111111111111111111111111111111111";
-        assert_eq!(rendered, expected);
-    }
-
-    #[test]
-    fn submitted_transaction_display__should_render_marker_without_metadata() {
-        // Given
-        let transaction = SubmittedTransaction {
-            metadata: None,
-            status: SubmittedTransactionStatus::SubmitFailed,
-            ..test_transaction("respond")
-        };
-
-        // When
-        let rendered = transaction.to_string();
-
-        // Then
-        let expected = "  2023-11-14 22:13:20.0 +00:00:00  SubmitFailed  method=respond                   signer=responder.near key=ed25519:US517G5965aydkZ46HS38QLi7UQiSojurfbQfKCELFx  (not submitted)";
-        assert_eq!(rendered, expected);
     }
 
     #[test]
